moved to plain webpack configuration bugfix unique id generation
