+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["react", "es2015"],
+  "presets": ["env", "es2015", "react"]
 }
