--- circle.yml
@@ -1,22 +1,0 @@
-machine:
-  pre:
-    - mkdir ~/.yarn-cache
-
-  node:
-    version: 6.9.1
-
-  environment:
-    YARN_VERSION: 0.18.1
-    PATH: "${PATH}:${HOME}/.yarn/bin:${HOME}/${CIRCLE_PROJECT_REPONAME}/node_modules/.bin"
-
-dependencies:
-  cache_directories:
-    - ~/.yarn-cache
-    - ~/.yarn
-
-  override:
-    - yarn install
-
-test:
-  override:
-    - yarn test
