+++ demo/src/index.js
@@ -3,7 +3,4 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import './index.css';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('demo')
-);
+ReactDOM.render(<App />, document.getElementById('root'));
