+++ package.json
@@ -3,60 +3,59 @@
   "version": "1.5.10",
   "type": "react-component",
   "description": "react-google-charts React component",
-  "main": "lib/index.js",
+  "main": "./dist/bundle.js",
   "scripts": {
-    "deploy": "gh-pages-deploy",
-    "build": "mv .babelrc .babelrc.backup && nwb build && mv .babelrc.backup .babelrc",
-    "clean": "nwb clean",
-    "start": "nwb serve",
-    "lint": "./node_modules/.bin/eslint src/*",
-    "test-node-env": "./node_modules/.bin/jest --env node __tests__/Chart.test.js  --verbose",
-    "test-jsdom-env": "./node_modules/.bin/jest --env jsdom __tests__/Chart.test.js --verbose",
     "test": "npm run lint && npm run test-node-env && npm run test-jsdom-env",
-    "gen:changelog": "github-changes -o RakanNimer -r react-google-charts"
+    "start": "webpack --watch",
+    "build": "webpack",
+    "lint": "eslint --fix src/*",
+    "test-node-env": "jest --env node __tests__/Chart.test.js  --verbose",
+    "test-jsdom-env": "jest --env jsdom __tests__/Chart.test.js --verbose",
+    "gen:changelog": "github-changes -o Lotti -r react-google-charts-apikey",
+    "prepublish": "npm run build"
   },
   "dependencies": {
     "debug": "^3.1.0",
-    "loadjs": "^3.3.1",
-    "prop-types": "^15.6.1",
+    "loadjs": "^3.5.4",
+    "prop-types": "^15.6.1"
+  },
+  "peerDependencies": {
     "react": "^16.3.2",
     "react-dom": "^16.3.2"
   },
-  "peerDependencies": {
-    "react": ">= 0.13.x",
-    "react-dom": ">= 0.13.x"
-  },
   "devDependencies": {
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "babel-core": "^6.26.3",
     "babel-eslint": "^8.2.3",
     "babel-jest": "^22.4.3",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "dotenv": "^5.0.1",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "clean-webpack-plugin": "^0.1.19",
     "eslint": "^4.19.1",
+    "eslint-cli": "^1.1.1",
     "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-babel": "^5.0.0",
+    "eslint-plugin-babel": "^5.1.0",
     "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.7.0",
-    "github-changes": "^1.0.4",
-    "gulp-react-docs": "^0.1.2",
     "jest": "^22.4.3",
-    "jsdoc": "^3.4.0",
-    "jsdom": "^11.8.0",
-    "nwb": "^0.21.5",
-    "object-assign": "^4.1.0",
     "pre-commit": "^1.2.2",
-    "react-router": "^4.2.0",
     "react-test-renderer": "^16.3.2",
-    "yargs": "^11.0.0"
+    "webpack": "^4.6.0",
+    "webpack-cli": "^2.0.15"
   },
   "gh-pages-deploy": {
     "staticpath": "demo/dist/"
   },
   "bugs": {
-    "url": "https://github.com/rakannimer/react-google-charts/issues"
+    "url": "https://github.com/lotti/react-google-charts-apikey/issues"
   },
   "jest": {
+    "browser": false,
+    "runner": "jest-runner",
     "moduleFileExtensions": [
       "jsx",
       "js",
@@ -72,10 +71,17 @@
     "testEnvironment": "node",
     "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(js|jsx)$"
   },
-  "author": "Rakan Nimer",
-  "homepage": "https://rakannimer.github.io/react-google-charts/",
+  "author": {
+    "name": "Rakan Nimer"
+  },
+  "contributors": [
+    {
+      "name": "Valerio \"Lotti\" Riva",
+      "email": "lotti86@gmail.com"
+    }
+  ],
   "license": "MIT",
-  "repository": "https://github.com/RakanNimer/react-google-charts",
+  "repository": "https://github.com/lotti/react-google-charts-apikey",
   "keywords": [
     "react",
     "google",
