+++ webpack.config.js
@@ -1,0 +1,55 @@
+const path = require('path');
+const CleanWebpackPlugin = require('clean-webpack-plugin');
+// Webpack configuration
+module.exports = {
+  mode: 'production',
+  context: path.join(__dirname, 'src'),
+  entry: './index',
+  output: {
+    path: path.join(__dirname, 'dist'),
+    filename: 'bundle.js',
+    library: 'react-google-charts-apikey',
+    libraryTarget: 'umd',
+    publicPath: '/dist/',
+    umdNamedDefine: true,
+  },
+  // Tell webpack to use html plugin
+  plugins: [
+    new CleanWebpackPlugin(['dist']),
+  ],
+  // Loaders configuration
+  // We are telling webpack to use "babel-loader" for .js and .jsx files
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: [
+          'babel-loader',
+        ],
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.js', '.jsx'],
+    alias: {
+      react: path.resolve(__dirname, './node_modules/react'),
+      'react-dom': path.resolve(__dirname, './node_modules/react-dom'),
+    },
+  },
+  externals: {
+    // Don't bundle react or react-dom
+    react: {
+      commonjs: 'react',
+      commonjs2: 'react',
+      amd: 'React',
+      root: 'React',
+    },
+    'react-dom': {
+      commonjs: 'react-dom',
+      commonjs2: 'react-dom',
+      amd: 'ReactDOM',
+      root: 'ReactDOM',
+    },
+  },
+};
